Attempting to use scrolled view
Issue hwere iws with the released event on the button
Looking at changing this to gestered tap and long hold gesture?

maybe tap and double tap?
Maui doesnt seem to like scrolling with buttons that are awaiting release
lose focus doesnt work as pulling focus from this isnt being triggered on a drag